tests/wpt/web-platform-tests/tools/wptrunner/test/testdata/testharness/testharness.https.html

<!doctype html>
<title>Example https test</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
  assert_equals(window.location.protocol, "https:");
}, "Test that file was loaded with the correct protocol");

</script>
